fix errors in treatment block (#3)
* fix error loggings to reflects issues better 

* robust element rendering: delay and retry until DOM element appears

* fallback function when render with retry fails 

* add user6, llm6, user7, llm7 to html for more robust fallback 

* documentation & linting

* set embedded variables`llm_response_*` before rendering DOM

The script finds the best place to put the messages which it does by looking at the innerhtml of all DOMs : LLM1_msg, LLM2_msg, LLM3_msg, etc. It puts the message in the first DOM that has LLMplaceholder as the innerhtml. Since you deleted LLM2_msg, it found that LLM3_msg had the innerhtml as LLMplaceholder and inserted the message there. What you can do to test this is go for four rounds, delete LLM5_msg and LLM5_dot and see what happens.

---------

Co-authored-by: Saumya Bhadani <saumya@pop-os.localdomain>
Co-authored-by: btrantruong <27907677+btrantruong@users.noreply.github.com>